Feat : HomeViewReactor 닉네임 최대 글자 로직 변경 - 닉네임 최대 글자 도달 시 사용자 인지를 위한 애니메이션 추가 - Mutation, State case 및 변수 추가 - 각 함수별 shakeNickNameTextField case 추가 - blockNickNameMaxLength() 함수에 needShake 로직 추가 및 Observable concat return
Issue #37
